tech-debt(shared): extract @straude/shared package for cross-package helpers
