+++ apps/web/app/api/embed/[username]/svg/route.ts
@@ -1,4 +1,5 @@
 import { NextRequest } from "next/server";
+import { formatTokens } from "@straude/shared/format";
 import { getServiceClient } from "@/lib/supabase/service";
 import { getProfileShareCardData } from "@/lib/share-assets/profile-card-data";
 import {
@@ -18,13 +19,6 @@ function esc(str: string): string {
     .replace(/</g, "&lt;")
     .replace(/>/g, "&gt;")
     .replace(/"/g, "&quot;");
-}
-
-function formatTokens(n: number): string {
-  if (n >= 1_000_000_000) return `${(n / 1_000_000_000).toFixed(1)}B`;
-  if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
-  if (n >= 1_000) return `${(n / 1_000).toFixed(1)}k`;
-  return String(n);
 }
 
 // ── SVG gradient workaround ────────────────────────────────────────
