+++ packages/cli/package.json
@@ -26,6 +26,7 @@
   },
   "dependencies": {
     "@pppp606/ink-chart": "^0.2.4",
+    "@straude/shared": "workspace:*",
     "chalk": "^5.6.2",
     "ink": "^6.8.0",
     "posthog-node": "^5.29.1",
