+++ packages/cli/src/components/PushSummary.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Box, Text } from 'ink';
+import { formatTokens } from '@straude/shared/format';
 import { theme } from './theme.js';
 import { BarChart } from './BarChart.js';
 import { ModelPalette } from './ModelPalette.js';
@@ -34,13 +35,6 @@ export interface PostResult {
 export interface PushSummaryProps {
   dashboard: DashboardData;
   results?: PostResult[];
-}
-
-function formatTokens(n: number): string {
-  if (n >= 1_000_000_000) return `${(n / 1_000_000_000).toFixed(1)}B`;
-  if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
-  if (n >= 1_000) return `${(n / 1_000).toFixed(1)}k`;
-  return String(n);
 }
 
 export function PushSummary({ dashboard, results }: PushSummaryProps) {
